requirements.txt
google-cloud-speech>=2.21.0
google-api-python-client>=2.120.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
fastapi>=0.111.0
uvicorn>=0.29.0
python-dotenv>=1.0.1
google-generativeai>=0.5.0
requests>=2.31.0
pytest>=8.2.0
flake8>=7.0.0
